requirements.txt
transformers>=4.45
accelerate 
torchaudio 
librosa 
soundfile
